Remove param utils
